pyproject.toml
[project]
name = "sam-converter"
version = "0.1.0"
description = "Convert SAM Designer T-SQL files to Snowflake SQL with dbt sources extraction"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlglot>=26.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
sam-converter = "sam_converter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.0"]
